Refactor examples and include TLS support
