+++ .travis.yml
@@ -17,7 +17,6 @@ jobs:
     - stage: test
       script:
         - go build ./...
-#        - go test ./...
         - go get golang.org/x/tools/cmd/cover
         - go get github.com/mattn/goveralls
         - go test ./ws ./ocppj -v -covermode=count -coverprofile=coverage.out
