+++ docker-compose.yml
@@ -1,24 +1,26 @@
 version: '3'
 services:
-  central_system:
+  central-system:
     build:
       context: .
       dockerfile: example/cs1.6/Dockerfile
     image: ldonini/ocpp1.6-central-system:latest
     container_name: central-system
+    environment:
+      - SERVER_LISTEN_PORT=8887
     ports:
       - "8887:8887"
     networks:
       - sim
-  charge_point:
-    environment:
-      - CLIENT_ID=chargePointSim
-      - CENTRAL_SYSTEM_URL=ws://central_system:8887
+  charge-point:
     build:
       context: .
       dockerfile: example/cp1.6/Dockerfile
     image: ldonini/ocpp1.6-charge-point:latest
     container_name: charge-point
+    environment:
+      - CLIENT_ID=chargePointSim
+      - CENTRAL_SYSTEM_URL=ws://central-system:8887
     networks:
       - sim
