+++ example/cp1.6/charge_point_sim.go
@@ -1,252 +1,68 @@
 package main
 
 import (
+	"crypto/tls"
+	"crypto/x509"
 	"fmt"
 	ocpp16 "github.com/lorenzodonini/ocpp-go/ocpp1.6"
 	"github.com/lorenzodonini/ocpp-go/ocpp1.6/core"
-	"github.com/lorenzodonini/ocpp-go/ocpp1.6/firmware"
 	"github.com/lorenzodonini/ocpp-go/ocpp1.6/localauth"
-	"github.com/lorenzodonini/ocpp-go/ocpp1.6/remotetrigger"
-	"github.com/lorenzodonini/ocpp-go/ocpp1.6/reservation"
-	"github.com/lorenzodonini/ocpp-go/ocpp1.6/smartcharging"
 	"github.com/lorenzodonini/ocpp-go/ocpp1.6/types"
+	"github.com/lorenzodonini/ocpp-go/ws"
 	log "github.com/sirupsen/logrus"
+	"io/ioutil"
 	"os"
+	"strconv"
 	"time"
 )
 
-type ConnectorInfo struct {
-	status             core.ChargePointStatus
-	availability       core.AvailabilityType
-	currentTransaction int
-	currentReservation int
+const (
+	envVarClientID             = "CLIENT_ID"
+	envVarCentralSystemUrl     = "CENTRAL_SYSTEM_URL"
+	envVarTls                  = "TLS_ENABLED"
+	envVarCACertificate        = "CA_CERTIFICATE_PATH"
+	envVarClientCertificate    = "CLIENT_CERTIFICATE_PATH"
+	envVarClientCertificateKey = "CLIENT_CERTIFICATE_KEY_PATH"
+)
+
+func setupChargePoint(chargePointID string) ocpp16.ChargePoint {
+	return ocpp16.NewChargePoint(chargePointID, nil, nil)
 }
 
-type ChargePointHandler struct {
-	status               core.ChargePointStatus
-	connectors           map[int]*ConnectorInfo
-	errorCode            core.ChargePointErrorCode
-	configuration        map[string]core.ConfigurationKey
-	meterValue           int
-	localAuthList        []localauth.AuthorizationData
-	localAuthListVersion int
-}
-
-var chargePoint ocpp16.ChargePoint
-
-// ------------- Core profile callbacks -------------
-
-func (handler *ChargePointHandler) OnChangeAvailability(request *core.ChangeAvailabilityRequest) (confirmation *core.ChangeAvailabilityConfirmation, err error) {
-	handler.connectors[request.ConnectorId].availability = request.Type
-	return core.NewChangeAvailabilityConfirmation(core.AvailabilityStatusAccepted), nil
-}
-
-func (handler *ChargePointHandler) OnChangeConfiguration(request *core.ChangeConfigurationRequest) (confirmation *core.ChangeConfigurationConfirmation, err error) {
-	configKey, ok := handler.configuration[request.Key]
-	if !ok {
-		return core.NewChangeConfigurationConfirmation(core.ConfigurationStatusNotSupported), nil
-	} else if configKey.Readonly {
-		return core.NewChangeConfigurationConfirmation(core.ConfigurationStatusRejected), nil
-	}
-	configKey.Value = request.Value
-	handler.configuration[request.Key] = configKey
-	return core.NewChangeConfigurationConfirmation(core.ConfigurationStatusAccepted), nil
-}
-
-func (handler *ChargePointHandler) OnClearCache(request *core.ClearCacheRequest) (confirmation *core.ClearCacheConfirmation, err error) {
-	return core.NewClearCacheConfirmation(core.ClearCacheStatusAccepted), nil
-}
-
-func (handler *ChargePointHandler) OnDataTransfer(request *core.DataTransferRequest) (confirmation *core.DataTransferConfirmation, err error) {
-	logDefault(request.GetFeatureName()).Infof("data transfer [Vendor: %v Message: %v]: %v", request.VendorId, request.MessageId, request.Data)
-	return core.NewDataTransferConfirmation(core.DataTransferStatusAccepted), nil
-}
-
-func (handler *ChargePointHandler) OnGetConfiguration(request *core.GetConfigurationRequest) (confirmation *core.GetConfigurationConfirmation, err error) {
-	var resultKeys []core.ConfigurationKey
-	var unknownKeys []string
-	for _, key := range request.Key {
-		configKey, ok := handler.configuration[key]
-		if !ok {
-			unknownKeys = append(unknownKeys, configKey.Value)
-		} else {
-			resultKeys = append(resultKeys, configKey)
-		}
-	}
-	conf := core.NewGetConfigurationConfirmation(resultKeys)
-	conf.UnknownKey = unknownKeys
-	return conf, nil
-}
-
-func (handler *ChargePointHandler) OnRemoteStartTransaction(request *core.RemoteStartTransactionRequest) (confirmation *core.RemoteStartTransactionConfirmation, err error) {
-	connector, ok := handler.connectors[request.ConnectorId]
-	if !ok {
-		return core.NewRemoteStartTransactionConfirmation(types.RemoteStartStopStatusRejected), nil
-	} else if connector.availability != core.AvailabilityTypeOperative || connector.status != core.ChargePointStatusAvailable || connector.currentTransaction > 0 {
-		return core.NewRemoteStartTransactionConfirmation(types.RemoteStartStopStatusRejected), nil
-	}
-	logDefault(request.GetFeatureName()).Infof("started transaction %v on connector %v", connector.currentTransaction, request.ConnectorId)
-	connector.currentTransaction = request.ConnectorId
-	return core.NewRemoteStartTransactionConfirmation(types.RemoteStartStopStatusAccepted), nil
-}
-
-func (handler *ChargePointHandler) OnRemoteStopTransaction(request *core.RemoteStopTransactionRequest) (confirmation *core.RemoteStopTransactionConfirmation, err error) {
-	for key, val := range handler.connectors {
-		if val.currentTransaction == request.TransactionId {
-			logDefault(request.GetFeatureName()).Infof("stopped transaction %v on connector %v", val.currentTransaction, key)
-			val.currentTransaction = 0
-			val.currentReservation = 0
-			val.status = core.ChargePointStatusAvailable
-			return core.NewRemoteStopTransactionConfirmation(types.RemoteStartStopStatusAccepted), nil
-		}
-	}
-	return core.NewRemoteStopTransactionConfirmation(types.RemoteStartStopStatusRejected), nil
-}
-
-func (handler *ChargePointHandler) OnReset(request *core.ResetRequest) (confirmation *core.ResetConfirmation, err error) {
-	//TODO: stop all ongoing transactions
-	return core.NewResetConfirmation(core.ResetStatusAccepted), nil
-}
-
-func (handler *ChargePointHandler) OnUnlockConnector(request *core.UnlockConnectorRequest) (confirmation *core.UnlockConnectorConfirmation, err error) {
-	_, ok := handler.connectors[request.ConnectorId]
-	if !ok {
-		return core.NewUnlockConnectorConfirmation(core.UnlockStatusNotSupported), nil
-	}
-	return core.NewUnlockConnectorConfirmation(core.UnlockStatusUnlocked), nil
-}
-
-// ------------- Local authorization list profile callbacks -------------
-
-func (handler *ChargePointHandler) OnGetLocalListVersion(request *localauth.GetLocalListVersionRequest) (confirmation *localauth.GetLocalListVersionConfirmation, err error) {
-	return localauth.NewGetLocalListVersionConfirmation(handler.localAuthListVersion), nil
-}
-
-func (handler *ChargePointHandler) OnSendLocalList(request *localauth.SendLocalListRequest) (confirmation *localauth.SendLocalListConfirmation, err error) {
-	if request.ListVersion <= handler.localAuthListVersion {
-		return localauth.NewSendLocalListConfirmation(localauth.UpdateStatusVersionMismatch), nil
-	}
-	if request.UpdateType == localauth.UpdateTypeFull {
-		handler.localAuthList = request.LocalAuthorizationList
-		handler.localAuthListVersion = request.ListVersion
-	} else if request.UpdateType == localauth.UpdateTypeDifferential {
-		handler.localAuthList = append(handler.localAuthList, request.LocalAuthorizationList...)
-		handler.localAuthListVersion = request.ListVersion
-	}
-	return localauth.NewSendLocalListConfirmation(localauth.UpdateStatusAccepted), nil
-}
-
-// ------------- Firmware management profile callbacks -------------
-
-func (handler *ChargePointHandler) OnGetDiagnostics(request *firmware.GetDiagnosticsRequest) (confirmation *firmware.GetDiagnosticsConfirmation, err error) {
-	return firmware.NewGetDiagnosticsConfirmation(), nil
-	//TODO: perform diagnostics upload out-of-band
-}
-
-func (handler *ChargePointHandler) OnUpdateFirmware(request *firmware.UpdateFirmwareRequest) (confirmation *firmware.UpdateFirmwareConfirmation, err error) {
-	return firmware.NewUpdateFirmwareConfirmation(), nil
-	//TODO: download new firmware out-of-band
-}
-
-// ------------- Remote trigger profile callbacks -------------
-
-func (handler *ChargePointHandler) OnTriggerMessage(request *remotetrigger.TriggerMessageRequest) (confirmation *remotetrigger.TriggerMessageConfirmation, err error) {
-	switch request.RequestedMessage {
-	case core.BootNotificationFeatureName:
-		//TODO: schedule boot notification message
-		break
-	case firmware.DiagnosticsStatusNotificationFeatureName:
-		//TODO: schedule diagnostics status notification message
-		break
-	case firmware.FirmwareStatusNotificationFeatureName:
-		//TODO: schedule firmware status notification message
-		break
-	case core.HeartbeatFeatureName:
-		//TODO: schedule heartbeat message
-		break
-	case core.MeterValuesFeatureName:
-		//TODO: schedule meter values message
-		break
-		//TODO: schedule status notification message
-	case core.StatusNotificationFeatureName:
-		break
-	default:
-		return remotetrigger.NewTriggerMessageConfirmation(remotetrigger.TriggerMessageStatusNotImplemented), nil
-	}
-	return remotetrigger.NewTriggerMessageConfirmation(remotetrigger.TriggerMessageStatusAccepted), nil
-}
-
-// ------------- Reservation profile callbacks -------------
-
-func (handler *ChargePointHandler) OnReserveNow(request *reservation.ReserveNowRequest) (confirmation *reservation.ReserveNowConfirmation, err error) {
-	connector := handler.connectors[request.ConnectorId]
-	if connector == nil {
-		return reservation.NewReserveNowConfirmation(reservation.ReservationStatusUnavailable), nil
-	} else if connector.status != core.ChargePointStatusAvailable {
-		return reservation.NewReserveNowConfirmation(reservation.ReservationStatusOccupied), nil
-	}
-	connector.currentReservation = request.ReservationId
-	go updateStatus(handler, request.ConnectorId, core.ChargePointStatusReserved)
-	// TODO: automatically remove reservation after expiryDate
-	return reservation.NewReserveNowConfirmation(reservation.ReservationStatusAccepted), nil
-}
-
-func (handler *ChargePointHandler) OnCancelReservation(request *reservation.CancelReservationRequest) (confirmation *reservation.CancelReservationConfirmation, err error) {
-	for k, v := range handler.connectors {
-		if v.currentReservation == request.ReservationId {
-			v.currentReservation = 0
-			if v.status == core.ChargePointStatusReserved {
-				go updateStatus(handler, k, core.ChargePointStatusAvailable)
-			}
-			return reservation.NewCancelReservationConfirmation(reservation.CancelReservationStatusAccepted), nil
-		}
-	}
-	return reservation.NewCancelReservationConfirmation(reservation.CancelReservationStatusRejected), nil
-}
-
-// ------------- Smart charging profile callbacks -------------
-
-func (handler *ChargePointHandler) OnSetChargingProfile(request *smartcharging.SetChargingProfileRequest) (confirmation *smartcharging.SetChargingProfileConfirmation, err error) {
-	//TODO: handle logic
-	return smartcharging.NewSetChargingProfileConfirmation(smartcharging.ChargingProfileStatusNotImplemented), nil
-}
-
-func (handler *ChargePointHandler) OnClearChargingProfile(request *smartcharging.ClearChargingProfileRequest) (confirmation *smartcharging.ClearChargingProfileConfirmation, err error) {
-	//TODO: handle logic
-	return smartcharging.NewClearChargingProfileConfirmation(smartcharging.ClearChargingProfileStatusUnknown), nil
-}
-
-func (handler *ChargePointHandler) OnGetCompositeSchedule(request *smartcharging.GetCompositeScheduleRequest) (confirmation *smartcharging.GetCompositeScheduleConfirmation, err error) {
-	//TODO: handle logic
-	return smartcharging.NewGetCompositeScheduleConfirmation(smartcharging.GetCompositeScheduleStatusRejected), nil
-}
-
-func checkError(err error) {
+func setupTlsChargePoint(chargePointID string) ocpp16.ChargePoint {
+	certPool, err := x509.SystemCertPool()
 	if err != nil {
 		log.Fatal(err)
 	}
-}
-
-func getExpiryDate(info *types.IdTagInfo) string {
-	if info.ExpiryDate != nil {
-		return fmt.Sprintf("authorized until %v", info.ExpiryDate.String())
+	// Load CA cert
+	caPath, ok := os.LookupEnv(envVarCACertificate)
+	if !ok {
+		log.Fatalf("no required %v found", envVarCACertificate)
 	}
-	return ""
-}
-
-func updateStatus(stateHandler *ChargePointHandler, connector int, status core.ChargePointStatus) {
-	if connector == 0 {
-		stateHandler.status = status
-	} else {
-		stateHandler.connectors[connector].status = status
+	caCert, err := ioutil.ReadFile(caPath)
+	if err != nil {
+		log.Warn(err)
+	} else if !certPool.AppendCertsFromPEM(caCert) {
+		log.Warn("no ca.cert file found, will use system CA certificates")
 	}
-	statusConfirmation, err := chargePoint.StatusNotification(connector, core.NoError, status)
-	checkError(err)
-	if connector == 0 {
-		logDefault(statusConfirmation.GetFeatureName()).Infof("status for all connectors updated to %v", status)
-	} else {
-		logDefault(statusConfirmation.GetFeatureName()).Infof("status for connector %v updated to %v", connector, status)
+	// Load client certificate
+	clientCertPath, ok1 := os.LookupEnv(envVarClientCertificate)
+	clientKeyPath, ok2 := os.LookupEnv(envVarClientCertificateKey)
+	var clientCertificates []tls.Certificate
+	if ok1 && ok2 {
+		certificate, err := tls.LoadX509KeyPair(clientCertPath, clientKeyPath)
+		if err == nil {
+			clientCertificates = []tls.Certificate{certificate}
+		} else {
+			log.Infof("couldn't load client TLS certificate: %v", err)
+		}
 	}
+	// Create client with TLS config
+	client := ws.NewTLSClient(&tls.Config{
+		RootCAs: certPool,
+		Certificates: clientCertificates,
+	})
+	return ocpp16.NewChargePoint(chargePointID, nil, client)
 }
 
 func exampleRoutine(chargePoint ocpp16.ChargePoint, stateHandler *ChargePointHandler) {
@@ -300,20 +116,27 @@ func exampleRoutine(chargePoint ocpp16.ChargePoint, stateHandler *ChargePointHan
 
 // Start function
 func main() {
-	// Parse arguments from env variables
-	id, ok := os.LookupEnv("CLIENT_ID")
+	// Load config
+	id, ok := os.LookupEnv(envVarClientID)
 	if !ok {
-		log.Print("Usage:\n\tocppClientId\n\tocppServerUrl")
+		log.Printf("no %v environment variable found, exiting...", envVarClientID)
 		return
 	}
-	csUrl, ok := os.LookupEnv("CENTRAL_SYSTEM_URL")
+	csUrl, ok := os.LookupEnv(envVarCentralSystemUrl)
 	if !ok {
-		log.Print("Usage:\n\tocppClientId\n\tocppServerUrl")
+		log.Printf("no %v environment variable found, exiting...", envVarCentralSystemUrl)
 		return
 	}
-	// Create a default OCPP 1.6 charge point
-	chargePoint = ocpp16.NewChargePoint(id, nil, nil)
-	// Set a handler for all callback functions
+	// Check if TLS enabled
+	t, _ := os.LookupEnv(envVarTls)
+	tlsEnabled, _ := strconv.ParseBool(t)
+	// Prepare OCPP 1.6 charge point (defined in handler.go)
+	if tlsEnabled {
+		chargePoint = setupTlsChargePoint(id)
+	} else {
+		chargePoint = setupChargePoint(id)
+	}
+	// Setup some basic state management
 	connectors := map[int]*ConnectorInfo{
 		1: {status: core.ChargePointStatusAvailable, availability: core.AvailabilityTypeOperative, currentTransaction: 0},
 	}
