+++ example/cs1.6/handler.go
@@ -1,0 +1,181 @@
+package main
+
+import (
+	"fmt"
+	"github.com/lorenzodonini/ocpp-go/ocpp1.6/core"
+	"github.com/lorenzodonini/ocpp-go/ocpp1.6/firmware"
+	"github.com/lorenzodonini/ocpp-go/ocpp1.6/types"
+	log "github.com/sirupsen/logrus"
+	"time"
+)
+
+var (
+	nextTransactionId = 0
+)
+
+//TODO: cache authorization
+
+// Charge Point state
+type TransactionInfo struct {
+	id          int
+	startTime   *types.DateTime
+	endTime     *types.DateTime
+	startMeter  int
+	endMeter    int
+	connectorId int
+	idTag       string
+}
+
+func (ti *TransactionInfo) hasTransactionEnded() bool {
+	return ti.endTime != nil && !ti.endTime.IsZero()
+}
+
+type ConnectorInfo struct {
+	status             core.ChargePointStatus
+	currentTransaction int
+}
+
+func (ci *ConnectorInfo) hasTransactionInProgress() bool {
+	return ci.currentTransaction >= 0
+}
+
+type ChargePointState struct {
+	status            core.ChargePointStatus
+	diagnosticsStatus firmware.DiagnosticsStatus
+	firmwareStatus    firmware.FirmwareStatus
+	connectors        map[int]*ConnectorInfo // No assumptions about the # of connectors
+	transactions      map[int]*TransactionInfo
+	errorCode         core.ChargePointErrorCode
+}
+
+func (cps *ChargePointState) getConnector(id int) *ConnectorInfo {
+	ci, ok := cps.connectors[id]
+	if !ok {
+		ci = &ConnectorInfo{currentTransaction: -1}
+		cps.connectors[id] = ci
+	}
+	return ci
+}
+
+type CentralSystemHandler struct {
+	chargePoints map[string]*ChargePointState
+}
+
+// ------------- Core profile callbacks -------------
+
+func (handler *CentralSystemHandler) OnAuthorize(chargePointId string, request *core.AuthorizeRequest) (confirmation *core.AuthorizeConfirmation, err error) {
+	logDefault(chargePointId, request.GetFeatureName()).Infof("client authorized")
+	return core.NewAuthorizationConfirmation(types.NewIdTagInfo(types.AuthorizationStatusAccepted)), nil
+}
+
+func (handler *CentralSystemHandler) OnBootNotification(chargePointId string, request *core.BootNotificationRequest) (confirmation *core.BootNotificationConfirmation, err error) {
+	logDefault(chargePointId, request.GetFeatureName()).Infof("boot confirmed")
+	return core.NewBootNotificationConfirmation(types.NewDateTime(time.Now()), defaultHeartbeatInterval, core.RegistrationStatusAccepted), nil
+}
+
+func (handler *CentralSystemHandler) OnDataTransfer(chargePointId string, request *core.DataTransferRequest) (confirmation *core.DataTransferConfirmation, err error) {
+	logDefault(chargePointId, request.GetFeatureName()).Infof("received data %d", request.Data)
+	return core.NewDataTransferConfirmation(core.DataTransferStatusAccepted), nil
+}
+
+func (handler *CentralSystemHandler) OnHeartbeat(chargePointId string, request *core.HeartbeatRequest) (confirmation *core.HeartbeatConfirmation, err error) {
+	return core.NewHeartbeatConfirmation(types.NewDateTime(time.Now())), nil
+}
+
+func (handler *CentralSystemHandler) OnMeterValues(chargePointId string, request *core.MeterValuesRequest) (confirmation *core.MeterValuesConfirmation, err error) {
+	logDefault(chargePointId, request.GetFeatureName()).Infof("received meter values for connector %v, transaction %v. Meter values:\n", request.ConnectorId, request.TransactionId)
+	for _, mv := range request.MeterValue {
+		logDefault(chargePointId, request.GetFeatureName()).Printf("%v", mv)
+	}
+	return core.NewMeterValuesConfirmation(), nil
+}
+
+func (handler *CentralSystemHandler) OnStatusNotification(chargePointId string, request *core.StatusNotificationRequest) (confirmation *core.StatusNotificationConfirmation, err error) {
+	info, ok := handler.chargePoints[chargePointId]
+	if !ok {
+		return nil, fmt.Errorf("unknown charge point %v", chargePointId)
+	}
+	info.errorCode = request.ErrorCode
+	if request.ConnectorId > 0 {
+		connectorInfo := info.getConnector(request.ConnectorId)
+		connectorInfo.status = request.Status
+		logDefault(chargePointId, request.GetFeatureName()).Infof("connector %v updated status to %v", request.ConnectorId, request.Status)
+	} else {
+		info.status = request.Status
+		logDefault(chargePointId, request.GetFeatureName()).Infof("all connectors updated status to %v", request.Status)
+	}
+	return core.NewStatusNotificationConfirmation(), nil
+}
+
+func (handler *CentralSystemHandler) OnStartTransaction(chargePointId string, request *core.StartTransactionRequest) (confirmation *core.StartTransactionConfirmation, err error) {
+	info, ok := handler.chargePoints[chargePointId]
+	if !ok {
+		return nil, fmt.Errorf("unknown charge point %v", chargePointId)
+	}
+	connector := info.getConnector(request.ConnectorId)
+	if connector.currentTransaction >= 0 {
+		return nil, fmt.Errorf("connector %v is currently busy with another transaction", request.ConnectorId)
+	}
+	transaction := &TransactionInfo{}
+	transaction.idTag = request.IdTag
+	transaction.connectorId = request.ConnectorId
+	transaction.startMeter = request.MeterStart
+	transaction.startTime = request.Timestamp
+	transaction.id = nextTransactionId
+	nextTransactionId += 1
+	connector.currentTransaction = transaction.id
+	info.transactions[transaction.id] = transaction
+	//TODO: check billable clients
+	logDefault(chargePointId, request.GetFeatureName()).Infof("started transaction %v for connector %v", transaction.id, transaction.connectorId)
+	return core.NewStartTransactionConfirmation(types.NewIdTagInfo(types.AuthorizationStatusAccepted), transaction.id), nil
+}
+
+func (handler *CentralSystemHandler) OnStopTransaction(chargePointId string, request *core.StopTransactionRequest) (confirmation *core.StopTransactionConfirmation, err error) {
+	info, ok := handler.chargePoints[chargePointId]
+	if !ok {
+		return nil, fmt.Errorf("unknown charge point %v", chargePointId)
+	}
+	transaction, ok := info.transactions[request.TransactionId]
+	if ok {
+		connector := info.getConnector(transaction.connectorId)
+		connector.currentTransaction = -1
+		transaction.endTime = request.Timestamp
+		transaction.endMeter = request.MeterStop
+		//TODO: meter data
+	}
+	logDefault(chargePointId, request.GetFeatureName()).Infof("stopped transaction %v - %v. Meter values:", request.TransactionId, request.Reason)
+	for _, mv := range request.TransactionData {
+		logDefault(chargePointId, request.GetFeatureName()).Printf("%v", mv)
+	}
+	return core.NewStopTransactionConfirmation(), nil
+}
+
+// ------------- Firmware management profile callbacks -------------
+
+func (handler *CentralSystemHandler) OnDiagnosticsStatusNotification(chargePointId string, request *firmware.DiagnosticsStatusNotificationRequest) (confirmation *firmware.DiagnosticsStatusNotificationConfirmation, err error) {
+	info, ok := handler.chargePoints[chargePointId]
+	if !ok {
+		return nil, fmt.Errorf("unknown charge point %v", chargePointId)
+	}
+	info.diagnosticsStatus = request.Status
+	logDefault(chargePointId, request.GetFeatureName()).Infof("updated diagnostics status to %v", request.Status)
+	return firmware.NewDiagnosticsStatusNotificationConfirmation(), nil
+}
+
+func (handler *CentralSystemHandler) OnFirmwareStatusNotification(chargePointId string, request *firmware.FirmwareStatusNotificationRequest) (confirmation *firmware.FirmwareStatusNotificationConfirmation, err error) {
+	info, ok := handler.chargePoints[chargePointId]
+	if !ok {
+		return nil, fmt.Errorf("unknown charge point %v", chargePointId)
+	}
+	info.firmwareStatus = request.Status
+	logDefault(chargePointId, request.GetFeatureName()).Infof("updated firmware status to %v", request.Status)
+	return &firmware.FirmwareStatusNotificationConfirmation{}, nil
+}
+
+// No callbacks for Local Auth management, Reservation, Remote trigger or Smart Charging profile on central system
+
+// Utility functions
+
+func logDefault(chargePointId string, feature string) *log.Entry {
+	return log.WithFields(log.Fields{"client": chargePointId, "message": feature})
+}
